[Code cleanup] Removed useless check in Main.loadWhiteList()
We checked for the presence of `tuxcraft-update.json` in new instance
directory, and do a fresh install if not found (with a huge "critical"
warning, but no means to stop).
If this file is lacking, this is a mistake from the update providers (or
the user manually deleted it, but... let's ignore that, not our fault if
they're stupid), and every user will be affected (and needs to be
delivered a patch).
Furthermore, the next update will also be considered a fresh install
since there will be no `tuxcraft-update.json` to validate the instance
as an 'official' TuxCraft modpack instance.
We shall consider this too harmful, and abort and show a popup.
For now an IOExcpetion's stacktrace is printed when caught and we still
proceed, but better handling and redirecting of errors informations is
scheduled, and this case will be at least treated as any other crash.

A todo has been put as a reminder of why an IOException could happen.

We may check again for this file later on, but in a different way or at
a different time than after copying the old instance, since aborting in
between leaves an unclean copied instance.