[Code cleanup/Gui.java] Added functions dedicated to popup warnings
This makes the code in NextButtonAction a bit more readable and will
allow other classes to trigger popup errors and warning in the future.
These functions logs the message to stderr in the same time.

As a reminder (but there will be later in the code, todos put) log
conventions used are:

[#]: error
[!]: warning
[=]: info
[+]: enabled
[-]: disabled
[*]: debug