[Fix] fix ugly aborts or unclear abort messages
Not aborting on missing `tuxcraft-update.json` in loadWhiteList() caused
two errors to be shwon, including one unclear. We now abort directly
with a clean message.

Throwing a RuntimeException without cause in popError(exception) made
the stacktrace unusable, we now have the full history of frames with
this exception, for easier local debugging when abort triggered by a
popError(exception) call.